move conversions into their own file
