+++ analyze/index.ts
@@ -1,8 +1,6 @@
 import type {
   ArcjetLogger,
   ArcjetRequestDetails,
-  SensitiveInfoEntity,
-  DetectedSensitiveInfoEntity,
   DetectSensitiveInfoResult,
   CustomDetect,
   SensitiveInfoConfig,
@@ -16,6 +14,11 @@ import type {
   BotType,
   EmailValidationResult,
 } from "./wasm/arcjet_analyze_js_req.component.js";
+import {
+  ConvertAnalyzeEntitiesToProtocolEntities,
+  ConvertDetectedSensitiveInfoEntityToAnalyzeEntity,
+  ConvertProtocolEntitiesToAnalyzeEntities,
+} from "./convert";
 
 import { wasm as componentCoreWasm } from "./wasm/arcjet_analyze_js_req.component.core.wasm?js";
 import { wasm as componentCore2Wasm } from "./wasm/arcjet_analyze_js_req.component.core2.wasm?js";
@@ -32,77 +35,6 @@ const FREE_EMAIL_PROVIDERS = [
 interface AnalyzeContext {
   log: ArcjetLogger;
   characteristics: string[];
-}
-
-export function ConvertProtocolEntitiesToAnalyzeEntities(
-  entity: SensitiveInfoEntity,
-): core.SensitiveInfoEntity {
-  if (entity === "email") {
-    return { tag: "email" };
-  }
-
-  if (entity === "phone-number") {
-    return { tag: "phone-number" };
-  }
-
-  if (entity === "ip-address") {
-    return { tag: "ip-address" };
-  }
-
-  if (entity === "credit-card-number") {
-    return { tag: "credit-card-number" };
-  }
-
-  return {
-    tag: "custom",
-    val: "custom",
-  };
-}
-
-export function ConvertDetectedSensitiveInfoEntityToAnalyzeEntity(
-  entity?: DetectedSensitiveInfoEntity,
-): core.SensitiveInfoEntity | undefined {
-  if (entity === "email") {
-    return { tag: "email" };
-  }
-
-  if (entity === "phone-number") {
-    return { tag: "phone-number" };
-  }
-
-  if (entity === "credit-card-number") {
-    return { tag: "credit-card-number" };
-  }
-
-  if (entity === "ip-address") {
-    return { tag: "ip-address" };
-  }
-
-  if (entity === "custom") {
-    return { tag: "custom", val: "custom" };
-  }
-}
-
-export function ConvertAnalyzeEntitiesToProtocolEntities(
-  entity: core.SensitiveInfoEntity,
-): DetectedSensitiveInfoEntity {
-  if (entity.tag === "email") {
-    return "email";
-  }
-
-  if (entity.tag === "ip-address") {
-    return "ip-address";
-  }
-
-  if (entity.tag === "credit-card-number") {
-    return "credit-card-number";
-  }
-
-  if (entity.tag === "phone-number") {
-    return "phone-number";
-  }
-
-  return "custom";
 }
 
 // TODO: Do we actually need this wasmCache or does `import` cache correctly?
